% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.nucleotide.chunk.counts.R
\name{get.nucleotide.chunk.counts}
\alias{get.nucleotide.chunk.counts}
\title{Get Nucleotide Chunk Counts}
\usage{
get.nucleotide.chunk.counts(key, chr, upstream = 1, downstream = 1,
  start = 1, end = -1)
}
\arguments{
\item{key}{List of specify trinucleotides to count}

\item{chr}{Chromosome}

\item{upstream}{Length upstream to read}

\item{downstream}{Length downstream to read}

\item{start}{Starting position}

\item{end}{Ending position}
}
\description{
Short description
}
\details{
Detailed description
}
\examples{
\dontrun{

}

}
\author{
\email{Fouad.Yousif@oicr.on.ca}
}
