% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.kataegis.R
\name{test.kataegis}
\alias{test.kataegis}
\title{Test Kataegis}
\usage{
test.kataegis(chromosome.num, somatic, units, npar = TRUE, print = TRUE,
  exprobntcx, output.name, ref.dir, chromosome.length.file)
}
\arguments{
\item{chromosome.num}{asdf}

\item{somatic}{asdf}

\item{units}{asdf}

\item{npar}{asdf}

\item{print}{asdf}

\item{exprobntcx}{asdf}

\item{output.name}{asdf}

\item{ref.dir}{Path to a directory containing the reference genome.}

\item{chromosome.length.file}{asdf}
}
\description{
Short description
}
\details{
Detailed description
}
\examples{
\dontrun{
test.kataegis(chromosome.num, somatic,units, npar=TRUE, print=TRUE, exprobntcx, output.name, ref.dir, chromosome.length.file)
}

}
\author{
\email{Fouad.Yousif@oicr.on.ca}
}
