% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.context.R
\name{get.context}
\alias{get.context}
\title{Get Context}
\usage{
get.context(file, start)
}
\arguments{
\item{file}{Reference files directory}

\item{start}{The position of the mutation gene}
}
\value{
The trinucleotide context.
}
\description{
Gets the 5' and 3' neighboring bases to the mutated base
}
\examples{
\dontrun{
get.context(file.path(referencie.genome.dir, 'chr1.fa'), c(158297133, 161176181))
}

}
\author{
Fouad Yousif

Fan Fan
}
